requirements.txt
Pillow~=9.4.0
pytesseract~=0.3.10
pandas~=1.5.3
numpy~=1.24.2
Unidecode~=1.3.6
opencv-python~=4.5.4.60
fastapi~=0.89.1
fitz~=0.0.1.dev2
pydantic~=1.10.5
uvicorn~=0.20.0
pypdf2~=3.0.1